backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
requests==2.32.5
python-multipart==0.0.21
python-dotenv==1.0.1
PyMuPDF==1.26.6
pymupdf4llm==0.2.9
python-pptx==0.6.21
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.0
langchain-qdrant==0.1.1
qdrant-client==1.11.0
fastembed==0.3.4
langgraph==0.2.26
tiktoken==0.7.0
numpy==1.26.4
